+++ module_1/exercises_vite_react/the-good-corner/frontend/src/components/ui/Logo.tsx
@@ -1,12 +1,13 @@
 import React from "react";
+import { Link } from "react-router-dom";
 
 const Logo: React.FC = () => {
   return (
     <h1>
-      <a href="/" className="button logo link-button">
+      <Link to="/" className="button logo link-button">
         <span className="mobile-short-label">TGC</span>
         <span className="desktop-long-label">THE GOOD CORNER</span>
-      </a>
+      </Link>
     </h1>
   );
 };