+++ module_1/exercises_vite_react/the-good-corner/frontend/src/config/router.tsx
@@ -1,7 +1,10 @@
 import App from "../App.tsx";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import Home from "../pages/Home.tsx";
+import Ad from "../pages/Ad.tsx";
+import Category from "../pages/Category.tsx";
 import NotFound from "../pages/NotFound.tsx";
+import AdForm from "../pages/AdForm.tsx";
 
 const router = createBrowserRouter([
   {
@@ -10,13 +13,25 @@ const router = createBrowserRouter([
     children: [
       {
         path: "/",
-        element: <Home />
+        element: <Home />,
       },
       {
-        path: "*",
-        element: <NotFound />,
+        path: "ads/:slug",
+        element: <Ad />,
+      },
+      {
+        path: "/post-ad",
+        element: <AdForm/>
+      },
+      {
+        path: "category/:slug",
+        element: <Category />,
       },
     ],
+  },
+  {
+    path: "*",
+    element: <NotFound />,
   },
 ]);