+++ module_1/exercises_vite_react/the-good-corner/frontend/src/pages/Ad.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { useParams } from "react-router-dom";
+
+const Ad: React.FC = () => {
+  const { slug } = useParams<{ slug: string }>();
+
+  return (
+    <div>
+      <div>
+        <h1>Annonce : {slug}</h1>
+        <p>Détails de l'annonce pour le slug : {slug}</p>
+      </div>
+    </div>
+  );
+};
+
+export default Ad;